refactor(layout_folder): change layour folder name
