+++ yaki_admin/src/features/customer/layouts/CustomerPageContentTeamList.vue
@@ -6,7 +6,7 @@ import {useTeamStore} from "@/stores/teamStore";
 import modalState from "@/features/modal/services/modalState";
 import {MODALMODE} from "@/constants/modalMode";
 
-import PageContentLayout from "@/global-layouts/PageContentLayout.vue";
+import PageContentLayout from "@/layouts/PageContentLayout.vue";
 import SideBarButton from "@/features/shared/components/SideBarButton.vue";
 import TeamWithCaptainCard from "@/features/customer/components/TeamWithCaptainCard.vue";
 import plusIcon from "@/assets/images/plus.png";
